requirements.txt
fastapi
uvicorn
langchain
langchain-huggingface
langserve
openai
python-dotenv
sse_starlette
httpx
pydantic
pydantic-settings
python-multipart
slowapi
structlog
pytest
pytest-asyncio
pytest-cov
psutil